requirements.txt
rocrate_lang @ git+https://github.com/Language-Research-Technology/rocrate-lang-py
requests==2.26
pandas==1.3.4
python-dotenv
pytest